+++ R/app_ui.R
@@ -63,7 +63,7 @@ app_ui <- function(request) {
                    sliderInput(inputId = "aux", min = 2, value = 2,
                                label = "Cantidad de Clusters", max = 10),
                    colourpicker::colourInput(
-                     "auxColor", NULL, value = "red", allowTransparent = T),
+                     "auxColor", NULL, value = "red", allowTransparent = TRUE),
                    radioSwitch("deleteNAaux", "eliminanaaux", c("eliminarai", "impsutar"))
           )
         )